+++ tests/compiler/compileProject.spec.ts
@@ -4,8 +4,8 @@ import {
   clean,
   compile,
   validate
-} from '../src/compiler/compileProject';
-import * as compileWorkersSrc from '../src/compiler/compileWorkers';
+} from '../../src/compiler/compileProject';
+import * as compileWorkersSrc from '../../src/compiler/compileWorkers';
 
 describe('compileProject.ts', () => {
   let cleanWorkers: sinon.SinonStub;
