+++ tests/glidelite/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig-base",
+  "compilerOptions": {
+    "types": ["node", "mocha", "chai"]
+  },
+  "include": ["**/*", "../../src/glidelite/**/*"]
+}
