+++ tests/tsconfig.json
@@ -1,7 +1,8 @@
 {
-  "extends": "../tsconfig-base",
-  "compilerOptions": {
-    "types": ["node", "mocha", "chai"]
-  },
-  "include": ["**/*", "../src/**/*"]
+  "files": [],
+  "include": [],
+  "references": [
+    { "path": "./compiler" },
+    { "path": "./glidelite" }
+  ]
 }
